backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
edge-tts
